requirements.txt
openai==1.76.0
python-dotenv==1.0.0
websockets==12.0
